requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.16
djangorestframework==3.11.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
PyMySQL==0.10.0
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
uritemplate==3.0.1
urllib3==1.26.2
gunicorn==20.0.4
pipenv==2020.11.15

